Tablas rol y genero funcionales

+++ Wiikinder/database/migrations/2021_11_30_100342_conjunto.php
@@ -15,7 +15,7 @@ class Conjunto extends Migration
     {
         Schema::create('conjuntos', function (Blueprint $table) {
             $table->string('correo');
-            $table->integer('id_rol');
+            $table->unsignedBigInteger('id_rol');
             $table->primary(['correo', 'id_rol']);
             $table->foreign('correo')->references('correo')->on('personas')->onDelete('cascade');
             $table->foreign('id_rol')->references('id')->on('rol')->onDelete('cascade');
